Support for pickle protocol + init() restrct
Added support to be able to send data over the pickle protocol to
graphite.
Commit contains the required tests and logic.
init() was changed to be able so that we can provide a init_type of
plaintext or pickle - This will then allow for plaintext_tcp and
plaintext_udp, etc.